Java calculate version 0.2

+++ src/Main.java
@@ -4,10 +4,10 @@ public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.print("Write first number: ");
-        int num1 = scanner.nextInt();
+        float num1 = scanner.nextFloat();
         System.out.print("Write second number: ");
-        int num2 = scanner.nextInt();
-        int res;
+        float num2 = scanner.nextFloat();
+        float res;
         System.out.print("Action: ");
         String action = scanner.nextLine();
         action = scanner.nextLine();
